// AngularVoteSystemApp/src/lib/smart-table/smart-table.type.ts
export interface IColumnFilter {
  lazy?: boolean;
  external?: boolean;
}

export interface IColumnFilterValue {
  [key: string]: any;
}

export interface IColumn {
  filter?: boolean | ((column: IItem, value: string) => boolean);
  label?: string;
  key: string;
  sorter?: boolean | ((a: IItem, b: IItem) => number);
  _style?: any;
  _props?: ITableHeaderCellProps;
  _classes?: string | string[] | Set<string> | { [klass: string]: any; };
}

export interface ITableHeaderCellProps {
  /**
   * Sets the color context of the component to one of CoreUI’s themed colors.
   *
   * @type 'primary' | 'secondary' | 'success' | 'danger' | 'warning' | 'info' | 'dark' | 'light' | string
   */
  color?: Colors;
  [key: string]: any;
}

export interface ISorter {
  resetable?: boolean;
  external?: boolean;
}

export interface IItem {
  _props?: ITableRowCellProps;
  _cellProps?: ITableCellProps;
  _selected?: boolean;
  [key: string]: number | string | any;
  // _deleted?: boolean;
}

export interface ITableCellProps {
  _all?: ITableRowCellProps;
  _attributes?: { [key: string]: any };
  [key: string]: ITableRowCellProps | undefined;
}

export interface IItemInternal extends IItem {
  _id?: number;
}

export interface ItemsPerPageSelect {
  external?: boolean;
  label?: string;
  values?: Array<number>;
}

// export interface ScopedColumns {
//   [key: string]: any
//   details?: (a: IItem, b: number) => ReactNode
// }

export interface ISorterValue {
  column?: string;
  state?: number | string;
}